поправил тест, под styled
